Potential fix for code scanning alert no. 17: Multiplication result converted to larger type
To fix the problem, we need to ensure that the multiplication is performed using a larger integer type to prevent overflow. We can achieve this by casting one of the operands to size_t before performing the multiplication. This will ensure that the multiplication is done using the larger type, and the result will not overflow.

  - Cast one of the operands (sa->block_count_hor or sa->block_count_ver) to size_t before performing the multiplication.
  - This change should be made on line 72 of the file libopenjpeg-2.3.1/src/lib/openjp2/sparse_array.c.

+++ libopenjpeg-2.3.1/src/lib/openjp2/sparse_array.c
@@ -69,7 +69,7 @@ opj_sparse_array_int32_t* opj_sparse_array_int32_create(OPJ_UINT32 width,
         return NULL;
     }
     sa->data_blocks = (OPJ_INT32**) opj_calloc(sizeof(OPJ_INT32*),
-                      sa->block_count_hor * sa->block_count_ver);
+                      (size_t)sa->block_count_hor * sa->block_count_ver);
     if (sa->data_blocks == NULL) {
         opj_free(sa);
         return NULL;
